@@ -1,24 +1,31 @@
 {
     "name": "hsbmaulana/phpcollection",
-    "version": "1.0.1",
+    "version": "2.0.0",
     "description": "Data Structure and Algorithm",
 
     "require": {
 
-        "php": "^7.2"
+        "php": "^8.0"
     },
 
     "require-dev": {
 
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^9.3"
     },
 
     "autoload": {
 
-        "psr-4": {
+        "psr-0": {
 
-            "Collections\\": "src/",
-            "Libraries\\": "lib/"
+            "Lists\\": "src/",
+            "Hashes\\": "src/",
+            "Trees\\": "src/",
+            "Graphs\\": "src/"
         }
+    },
+
+    "autoload-dev": {
+
+        "classmap": [ "test/" ]
     }
 }
